+++ src/core/Codend.Domain/Core/Errors/DomainErrors.cs
@@ -30,42 +30,4 @@ public static partial class DomainErrors
             }
         }
     }
-    
-    public static class StoryDescription
-    {
-        public class NullOrEmpty : DomainError
-        {
-            public NullOrEmpty()
-                : base("StoryDescription.NullOrEmpty", "Story description can't be null or empty.")
-            {
-            }
-        }
-
-        public class DescriptionTooLong : DomainError
-        {
-            public DescriptionTooLong()
-                : base("StoryDescription.DescriptionTooLong", "Story description is too long.")
-            {
-            }
-        }
-    }
-
-    public static class StoryName
-    {
-        public class NullOrEmpty : DomainError
-        {
-            public NullOrEmpty()
-                : base("StoryName.NullOrEmpty", "Story name can't be null or empty.")
-            {
-            }
-        }
-
-        public class NameTooLong : DomainError
-        {
-            public NameTooLong()
-                : base("StoryName.DescriptionTooLong", "Story name is too long.")
-            {
-            }
-        }
-    }
 }
